// src/app/shared/model/app-ins-obj-obj.model.ts
export class AppInsObjObj {
    AppInsObjId: number;
    AppId: number;
    AppAssetId: number;
    AppCollateralId: number;
    AgrmntId: number;
    InsObjName: string;
    InscoBranchCode: string;
    InscoBranchName: string;
    InsSeqNo: number;
    MrInsRateTypeCode: string;
    InsAssetCoveredByCode: string;
    InsAssetCoverPeriod: string;
    InsAssetCoverPeriodName: string;
    InsAssetCoveredBy: string;
    InsAssetCoveredByName: string;
    InsAssetPaidBy: string;
    InsAssetPaidByName: string;
    TotalCustMainPremiAmt: number;
    TotalCustAddPremiAmt: number;
    TotalInscoMainPremiAmt: number;
    TotalInscoAddPremiAmt: number;
    CustAdminFeeAmt: number;
    InscoAdminFeeAmt: number;
    InscoStampDutyFee : number;
    CustStampDutyFee : number;
    Notes: string;
    InsStat: string;
    CustCoverStartDt: Date;
    InsStep: string;
    StartDt: Date;
    EndDt: Date;
    InsLength: number;
    MrInsPayMethodCode: string;
    CvgAmt: number;
    InsCpltzAmt: number;
    IsAdditionalIns: boolean;
    InsAssetRegion: string;
    InsPolicyNo: string;
    InsPolicyName: string;
    TotalCustDiscAmt: number;
    TotalInscoDiscAmt: number;
    TotalInsCustAmt: number;
    TotalInsInscoAmt: number;
    CustInscoBranchName: string;
    CustCvgAmt: number;
    CustNotes: string;
    PayPeriodToInsco: string;
    RowVersion: string;
    CustTotalFee: number;
    InscoTotalFee: number;

constructor() { 
    this.AppInsObjId = 0;
    this.AppId = 0;
    this.AppAssetId = null;
    this.AppCollateralId = null;
    this.AgrmntId = null;
    this.InsObjName = "";
    this.InscoBranchCode = "";
    this.InscoBranchName = "";
    this.InsSeqNo = 0;
    this.MrInsRateTypeCode = "";
    this.InsAssetCoveredByCode = "";
    this.InsAssetCoverPeriod = "-";
    this.InsAssetCoverPeriodName = "";
    this.InsAssetCoveredBy = "";
    this.InsAssetCoveredByName = "";
    this.InsAssetPaidBy = "";
    this.InsAssetPaidByName = "";
    this.TotalCustMainPremiAmt = 0;
    this.TotalCustAddPremiAmt = 0;
    this.TotalInscoMainPremiAmt = 0;
    this.TotalInscoAddPremiAmt = 0;
    this.CustAdminFeeAmt= 0;
    this.InscoAdminFeeAmt = 0;
    this.CustStampDutyFee = 0;
    this.InscoStampDutyFee = 0;
    this.Notes = "";
    this.InsStat = "";
    this.CustCoverStartDt = null;
    this.InsStep = "";
    this.StartDt = null;
    this.EndDt = null;
    this.InsLength = 0;
    this.MrInsPayMethodCode = "";
    this.CvgAmt = 0;
    this.InsCpltzAmt = 0;
    this.IsAdditionalIns = false;
    this.InsAssetRegion = "";
    this.InsPolicyNo = "";
    this.InsPolicyName = "";
    this.TotalCustDiscAmt = 0;
    this.TotalInscoDiscAmt = 0;
    this.TotalInsCustAmt = 0;
    this.TotalInsInscoAmt = 0;
    this.CustCvgAmt = 0;
    this.CustInscoBranchName = "";
    this.CustNotes = "";
    this.RowVersion = "";
    this.CustTotalFee = 0;
    this.InscoTotalFee = 0;
    }
}